Add `Table::iter_bytes`, `Table::scan_bytes` and `Table::size_bytes`
This allows for iterating over the raw, serialized bytes for each entry
in the table. The `Table::iter` and `Table::scan` implementations have
been simplified by using these new methods internally.

This allows for the easy calculation of the size of a table in bytes,
implemented as `Table::size_bytes`. A similar function called
`tree_size_bytes` has been added to the root for calculating the size of
a full `sled::Tree`.